Fix EH callsite table in basic-block-sections mode.
When exception support was added for basic-block-sections in
8955950c121c97a686310991203c89ba14c90b82, the commit said:
> The CallSiteTableLength will not actually represent the length of
> the call site table, but rather the reference to the action
> table. Since the only purpose of this field is to locate the action
> table, correctness is guaranteed.

This is not precisely correct: the unwind library will iterate over
callsite entries until either the length has been reached, a callsite
entry has matched the lookup pointer, or the next entry has a start
address greater-or-equal to that pointer.

Thus, if the end of the function is not covered by an entry, and the
length is too large, it will iterate past the "real" end, into
garbage.

In LLVM, today, addresses not covered by the callsite table may occur
only for 'nounwind' calls. As such calls should not unwind, this issue
is "mostly harmless" -- it causes failure only when you're doing
something bad already. But, in the future, when gaps are used to
implement 'unwindabort' behavior, this becomes more serious.

To address this, while keeping the ability to share the action table,
we ensure there is always be an callsite table entry covering the end
of the basic-block-section, when an invalid length is used.

An alternative would be to stop sharing the action table at all --
this would entirely eliminate the hack of using an invalid table
length, which might be a better idea...